requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0  # For YOLOv8
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0  # For CLIP and NLP models
faiss-cpu>=1.7.4  # For similarity search
Pillow>=10.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
spacy>=3.6.0
tqdm>=4.65.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0 